Add compatibility with the SmartPunctExtension
This duplicates most of the code from the SmartPunctExtension
in order to make it work for LaTeX-style quotes and ellipses
and to remove its handling of dashes.

Ideally most of these classes would be able to extend their
SmartPunctExtension counterparts, but those are `final` and so
this is the hacky workaround.

Refs #12